Drop the last deposit wizard concern
The wizard's controller had one remaining concern left over from the
refactor, used only by the controller itself. Folding its handful of
methods back into the controller leaves the wizard with no concerns
at all — just the controller, the presenter, and its value objects.
